% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{abonnement}
\alias{abonnement}
\title{Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre}
\usage{
abonnement(rnwFil, brukernavn = "tullebukk", reshID = 0)
}
\arguments{
\item{rnwFil}{Navn på fila som skal kjøres. Angis MED filending (\emph{dvs "filnavn.Rnw"})}

\item{brukernavn}{}

\item{reshID}{Aktuell reshid}
}
\value{
Full path of file produced
}
\description{
Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre
}
