% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRHighcharts.R
\name{AndelerHighchart}
\alias{AndelerHighchart}
\title{Test of Highchart for intensiv}
\usage{
AndelerHighchart(outfile)
}
\arguments{
\item{outfile}{String providing fully qualified path and name of html widget}
}
\description{
asæglksa
}

