% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{AndelerTab}
\alias{AndelerTab}
\title{Tabell med andel av div. variabler for koblet datasett (intensiv+beredskap)
Kun ferdigstilte registreringer}
\usage{
AndelerTab(
  RegData,
  datoFra = "2020-01-01",
  datoTil = Sys.Date(),
  erMann = 9,
  dodInt = 9,
  valgtRHF = "Alle",
  resp = 9,
  minald = 0,
  maxald = 110
)
}
\arguments{
\item{valgtRHF}{'Alle' (standard), RHF-navn uten 'Helse '}
}
\value{
andel for ulike nøkkelverdier
}
\description{
Tabell med andel av div. variabler for koblet datasett (intensiv+beredskap)
Kun ferdigstilte registreringer
}
