% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andelerTableContainer.R
\name{AndelerTableContainer}
\alias{AndelerTableContainer}
\title{Tabel container for report in shiny}
\usage{
AndelerTableContainer(groupText = "Gruppe", deptName = "ShusAvd")
}
\arguments{
\item{groupText}{String}

\item{deptName}{String}
}
\description{
Framework for table layout
}

