% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppNIRbrukervalg.R
\name{AppNIRbrukervalg}
\alias{AppNIRbrukervalg}
\title{Funksjon som inneholder alle brukervalg for Intensivregisteret}
\usage{
AppNIRbrukervalg(ark = "angi sideID")
}
\arguments{
\item{ark}{Hvilken side i appen skal vises? (fordelinger eller andeler)}
}
\value{
Visning av brukervalg
}
\description{
Denne funksjonen beregner ikke noe. :-|
}
\details{
Mer info?
}
