% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjonerPaaror.R
\name{beregneSkaarer}
\alias{beregneSkaarer}
\title{Fil som inneholder hjelpefunksjoner.}
\usage{
beregneSkaarer(RegData)
}
\arguments{
\item{RegData}{data}
}
\description{
Beregner skårer fra pårørendeskjema (FS-ICU). NB: Ikke tatt høyde for at 
minst 70% av spm. må være fylt ut for at en skår skal beregnes
}
