% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/reportObjects.R
\name{emptyReport}
\alias{emptyReport}
\title{Provider of report objects}
\usage{
emptyReport(Tittel, utvalg = "", infoText = "Tomt...")
}
\arguments{
\item{Tittel}{String Title to show in empty chart}

\item{infoText}{String Text to shown in empty chart}
}
\value{
plotObj Highchart object

tableObj Table object
}
\description{
Provide report object when no data (test replacement of NIRHighcharts)
}

