% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{finnDblReg}
\alias{finnDblReg}
\title{Finn eventuelle dobbeltregistreringer}
\usage{
finnDblReg(RegData, datoTil = Sys.Date(), reshID = 0, pasientID = "PasientID")
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato}

\item{reshID}{enhetens resh}

\item{pasientID}{pasientidentifikasjon, personentydig}
}
\description{
Finn eventuelle dobbeltregistreringer
}
