% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{henteSamleRapporter}
\alias{henteSamleRapporter}
\title{Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and 
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre}
\usage{
henteSamleRapporter(filnavn, rnwFil, reshID = 0, datoFra = startDato,
  datoTil = Sys.Date())
}
\arguments{
\item{reshID}{Aktuell reshid}

\item{registryName}{Registernavn - fjerne?}

\item{hospitalName}{fjerne?}

\item{type}{Single element list which value defining report file format,
currently one of "pdf" or "html"}
}
\value{
Full path of file produced
}
\description{
Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and 
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre
}
