% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{hjelpeFunksjoner}
\alias{hjelpeFunksjoner}
\alias{FinnReinnleggelser}
\alias{SorterOgNavngiTidsEnhet}
\alias{lageTulleData}
\alias{leggTilIntervensjon}
\alias{delTekst}
\alias{KobleMedHoved}
\title{Hjelpefunksjoner. Group of functions page title}
\usage{
FinnReinnleggelser(RegData, PasientID = "PasientID")

SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Aar", tab = 0)

lageTulleData(RegData, varBort = "", antSh = 26, antObs = 20000)

leggTilIntervensjon(RegData, startDatoIntervensjon = "2016-10-01",
  sluttDatoIntervensjon = Sys.Date())

delTekst(x, len)

KobleMedHoved(HovedSkjema, Skjema2, alleHovedskjema = F,
  alleSkjema2 = F)
}
\arguments{
\item{RegData}{data}

\item{PasientID}{Variabelen som angir pasientidentifikasjon}

\item{x}{En tekststreng eller vektor av tekststrenger}

\item{len}{Lengden strengen skal brytes ved}

\item{HovedSkjema}{Registerets hovedskjma (Main..)}

\item{Skjema2}{Skjemaet som skal kobles til hovedskjema. (Pårørendeskjema, Influensaskjema)}

\item{alleHovedskjema}{TRUE/FALSE. standard: FALSE. I praksis om vi skal ha en left? join eller ikke}

\item{alleSkjema2}{TRUE/FALSE.standard: FALSE I praksis om vi skal ha en right? join eller ikke}
}
\value{
Div hjelpefunksjoner
}
\description{
Fil med div hjelpefunksjoner.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.
}
\section{Finne reinnleggelser After function section}{

Despite its location, this actually comes after the function section.
Fil som inneholder hjelpefunksjoner. 
FinnReinnleggelser beregner reinnleggelser fra DateAdmittedIntensive og DateDischargedIntensive
SorterOgNavngiTidsEnhet Legger til tidsenhetene Aar, Halvaar, Mnd og Kvartal
}

\section{Tilrettelegge tidsenhetvariabel}{

Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.
}

\section{Lage tulledata (simulerte data)}{
NA
}

\section{Legge til indikator for intervensjon, pårørendeoppfølging}{
NA

NA
}

\section{Automatisk linjebryting av lange tekstetiketter}{
NA
}

