% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{kjorIntensivApp}
\alias{kjorIntensivApp}
\title{Kjøre Shiny Application}
\usage{
kjorIntensivApp(browser = FALSE, logAsJson = FALSE)
}
\arguments{
\item{browser}{TRUE/FALSE}

\item{logAsJson}{noe med logging...}
}
\value{
Et objekt som representerer den Rygg-app'en
}
\description{
Kjøre Shiny Application
}
