% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{KobleMedHoved}
\alias{KobleMedHoved}
\title{Legge til indikator for intervensjon, pårørendeoppfølging}
\usage{
KobleMedHoved(HovedSkjema, Skjema2, alleHovedskjema = F,
  alleSkjema2 = F)
}
\arguments{
\item{HovedSkjema}{Registerets hovedskjma (Main..)}

\item{Skjema2}{Skjemaet som skal kobles til hovedskjema. (Pårørendeskjema, Influensaskjema)}

\item{alleHovedskjema}{TRUE/FALSE. standard: FALSE. I praksis om vi skal ha en left? join eller ikke}

\item{alleSkjema2}{TRUE/FALSE.standard: FALSE I praksis om vi skal ha en right? join eller ikke}
}
\description{
Legge til indikator for intervensjon, pårørendeoppfølging
}
