% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{lageTulleData}
\alias{lageTulleData}
\title{Lage tulledata (simulerte data)}
\usage{
lageTulleData(RegData, varBort = "", antSh = 26, antObs = 20000)
}
\arguments{
\item{RegData}{Dataramme}

\item{varBort}{variable som skal fjernes fra RegData}

\item{antSh}{antall simulerte sykehus}

\item{antObs}{antall rader i det fiktive datasettet}
}
\description{
Lage tulledata (simulerte data)
}
