% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{lagTabavFig}
\alias{lagTabavFig}
\title{Vise figurdata som tabell}
\usage{
lagTabavFig(UtDataFraFig)
}
\arguments{
\item{UtDataFraFig}{data fra figurfunksjoner, dvs. beregnede verdier}
}
\description{
Vise figurdata som tabell
}
