% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{leggTilIntervensjon}
\alias{leggTilIntervensjon}
\title{Legge til indikator for intervensjon, pårørendeoppfølging}
\usage{
leggTilIntervensjon(
  RegData,
  startDatoIntervensjon = "2016-10-01",
  sluttDatoIntervensjon = Sys.Date()
)
}
\arguments{
\item{RegData}{Dataramme}

\item{startDatoIntervensjon}{startdato for intervensjon. Foreslått verdi '2016-01-01' basert 
på de første studiene som ble gjort med pårørendeskjema}

\item{sluttDatoIntervensjon}{sluttdato for}
}
\description{
Legge til indikator for intervensjon, pårørendeoppfølging
}
