% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRAndeler.R
\name{NIRAndeler}
\alias{NIRAndeler}
\title{Søylediagram som viser andeler av ulike variabler}
\usage{
NIRAndeler(RegData, valgtVar, datoFra = "2011-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, erMann = "",
  InnMaate = "", dodInt = "", outfile = "", preprosess = 1,
  hentData = 0, reshID, enhetsUtvalg = 1)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{InnMaate}{0: Elektivt, 
6: Akutt medisinsk, 
8: Akutt kirurgisk, 
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{dodInt}{Levende/død ut fra intensiv. 
0: i live, 
1: død,   
alle, standard (alle andre verdier)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{overfPas}{Overført under pågående intensivbehandling? 
1 = Pasienten er ikke overført
2 = Pasienten er overført}
}
\value{
Søylediagram (fordeling) av valgt variabel. De enkelte verdiene kan også sendes med.
}
\description{
Denne funksjonen beregner andeler (fordeling) av valgt variabel
filtrert på de utvalg som er gjort. Kan trenge funksjonerne:
NIRUtvalgEnh (skal endre navn til NIRUtvalg når ferdig)
NIRFigSoyler
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Aldersfordeling, 10-årige grupper 
    \item InnMaate: Hastegrad inn på intensiv (Elektivt, Akutt medisinsk, Akutt kirurgisk)
    \item liggetid: Liggetid 
    \item NEMS: Skår for ressursbruk. (Nine Equivalents of Nursing Manpower Use Score)
    \item Nas: Skår for sykepleieraktiviteter. (Nursing Activities Score)
    \item respiratortid: Tid tilbrakt i respirator
    \item SAPSII: Skår for alvorlighetsgrad av sykdom.  (Simplified Acute Physiology Score II)
   }
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
   \itemize{
    \item 0: Hele landet
    \item 1: Egen enhet mot resten av landet (Standard)
    \item 2: Egen enhet
    \item 3: Egen enhet mot egen sykehustype
    \item 4: Egen sykehustype
    \item 5: Egen sykehustype mot resten av landet
    \item 6: Egen enhet mot egen region [NB: Intensivregiisteret mangler pt. variabel for region]
    \item 7: Egen region [NB: Mangler pt. variabel for region]
   \item 8: Egen region mot resten [NB: Mangler pt. variabel for region]
   	}
}

