% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRAndelerGrVar.R
\name{NIRAndelerGrVar}
\alias{NIRAndelerGrVar}
\title{Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NIRAndelerGrVar(RegData, valgtVar, minald = 0, maxald = 130,
  datoFra = "2011-01-01", datoTil = "3000-01-01", grType = 99,
  grVar = "", InnMaate = 99, dodInt = "", erMann = "", hentData = 0,
  preprosess = 1, outfile = "", lagFig = 1)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{grType}{Gjør gruppeutvalg for
 1 eller 2: lokal-/sentralsykehus
3: regionsykehus}

\item{InnMaate}{0: Elektivt, 
6: Akutt medisinsk, 
8: Akutt kirurgisk, 
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{dodInt}{Levende/død ut fra intensiv. 
0: i live, 
1: død,   
alle, standard (alle andre verdier)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}
}
\value{
Søylediagram med andeler av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med andeler av en variabel for en grupperingsvariabelen sykehus.
Funksjonen er tilrettelagt for å kunne brukes til andre grupperingsvariable enn sykehus
}
\details{
Figurtypen som genereres er avhengig av valgtVar. Ved valgtVar='InnMaate', får man ei oversikt oversikt
over andel av innkomsttype ved hver enhet.
For valgtVar lik reinn eller DodeInt viser figuren andel av hhv. reinnleggelser og døde for hver enhet.
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder_u18: Pasienter under 18 år 
    \item alder_over80: Pasienter over 80 år (>=80)
    \item dod30d: Pasienter som dør innen 30 dager etter innleggelse
    \item dodeIntensiv: Pasienter som dør på intensivavdelinga. 
    \item innMaate: Hastegrad inn på intensiv (Elektivt, Akutt medisinsk, Akutt kirurgisk)
	Dette valget viser en annen figurtype.
    \item respStotte: Pasienter som har fått respiratorstøtte
    \item reinn: Andel reinnlagte (kun hvor dette er registrert, dvs. fjerner ukjente)
   }
}

