% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRAndelerGrVarOff.R
\name{NIRAndelerGrVarOff}
\alias{NIRAndelerGrVarOff}
\title{Søylediagram med AggVerdier for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NIRAndelerGrVarOff(RegData, valgtVar = "reinn", aar = 0, grType = 99,
  grVar = "ShNavn", InnMaate = 99, erMann = "", aldGr = 0,
  hentData = 0, outfile = "", lagFig = 0, utvalgsInfo = "",
  tittel = "", KImaal = 0)
}
\arguments{
\item{RegData}{Anonymisert datasett}

\item{valgtVar}{Kvalitetsindikatoren det vises resultat for. Angir også i hvilken fil RegData er lagret.}

\item{grType}{Gjør gruppeutvalg på sykehustype
1: lokal-/sentralsykehus
2: lokal-/sentralsykehus
3: regionsykehus
99: alle (standard)}

\item{InnMaate}{0: Elektivt, 
6: Akutt medisinsk, 
8: Akutt kirurgisk, 
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hentData}{Angir om data skal hentes fra pakken. Filnavn er gitt ved 
Filnavn gitt ved 'RegData01' + valgtVar + '.RData'}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{lagFig}{Angir om figur skal lages eller ikke 0-ikke lag, 1-lag}
}
\value{
Søylediagram med AggVerdier av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med AggVerdier av en gitt variabel for ei valgt gruppering, 
f.eks. enheter. I øyeblikket benytter funksjonen bare 'ShNavn' som grupperingsvariabel, men 
andre valg kan lett inkluderes.
}
\details{
Funksjonen benytter funksjonene: NIRRegDataSQL, NIRPreprosess, NIRVarTilrettelegg, NIRUtvalgEnh
og NIRFigSoyler.
Figurtypen som genereres er avhengig av valgtVar. Ved valgtVar='InnMaate', får man ei oversikt oversikt
over andel av innkomsttype ved hver enhet, dvs. ei fordeling av de tre innkomsttypene for hver enhet.
For andre "valgtVar" viser figuren andel av den valgte variabelen for hver enhet.
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder_u18: Pasienter under 18 år 
    \item alder_over80: Pasienter over 80 år (>=80)
    \item dod30d: Pasienter som dør innen 30 dager etter innleggelse
    \item dodeIntensiv: Pasienter som dør på intensivavdelinga. 
    \item innMaate: Hastegrad inn på intensiv (Elektivt, Akutt medisinsk, Akutt kirurgisk)
	Dette valget viser en annen figurtype.
    \item respStotte: Pasienter som har fått respiratorstøtte
    \item reinn: Andel reinnlagte (fjerner ukjente) Kvalitetsindikator
   } valgtVar vil her bestemme hvilken vedlagt datafil som skal benyttes
}
