% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigAndelerGrVar.R
\name{NIRFigAndelerGrVar}
\alias{NIRFigAndelerGrVar}
\title{Søylediagram med AggVerdier for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NIRFigAndelerGrVar(RegData, valgtVar, datoFra = "2010-01-01",
  datoTil = "3000-01-01", aar = 0, minald = 0, maxald = 130,
  aldGr = 0, medKI = 0, grType = 99, grVar = "ShNavn", InnMaate = 99,
  dodInt = "", erMann = "", hentData = 0, preprosess = 1,
  outfile = "", lagFig = 1, offData = 0)
}
\arguments{
\item{aldGr:}{Aldersgrupper. Brukes i offentliggjøringsfigurer}

\item{tittel:}{Hvis vil angi tittel direkte}

\item{utvalgsInfo:}{Hvis datafil lagret med utvalgsinfo}

\item{offData:}{Hvis vi benytter anonymiserte 01-data til offentliggjøring.}

\item{sortAvtagende:}{sortere søylene i figuren avtagende eller stigende.}
}
\value{
Søylediagram med AggVerdier av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med AggVerdier av en gitt variabel for ei valgt gruppering, 
f.eks. enheter. I øyeblikket benytter funksjonen bare 'ShNavn' som grupperingsvariabel, men 
andre valg kan lett inkluderes. Funksjonen er tilrettelagt for også å kunne benytte "01-data", 
dvs. kodede, anonymiserte data som benyttes til offentlige kvalitetsindikatorer.
}
\details{
Funksjonen benytter funksjonene: NIRRegDataSQL, NIRPreprosess, NIRVarTilrettelegg, NIRUtvalgEnh
og NIRFigSoyler.
Figurtypen som genereres er avhengig av valgtVar. Ved valgtVar='InnMaate', får man ei oversikt oversikt
over andel av innkomsttype ved hver enhet, dvs. ei fordeling av de tre innkomsttypene for hver enhet.
For andre "valgtVar" viser figuren andel av den valgte variabelen for hver enhet.
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder_u18: Pasienter under 18 år 
    \item alder_over80: Pasienter over 80 år (>=80)
    \item dod30d: Pasienter som dør innen 30 dager etter innleggelse
    \item dodeIntensiv: Pasienter som dør på intensivavdelinga. 
    \item innMaate: Hastegrad inn på intensiv (Elektivt, Akutt medisinsk, Akutt kirurgisk)
	Dette valget viser en annen figurtype.
    \item respStotte: Pasienter som har fått respiratorstøtte
    \item reinn: Andel reinnlagte (fjerner ukjente) Kvalitetsindikator
    \item Trakeostomi: Andel som har fått trakeostomi (kat 2 og 3)
   }
}
