% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigGjsnGrVar.R
\name{NIRFigGjsnGrVar}
\alias{NIRFigGjsnGrVar}
\title{Søylediagram med gjennomsnitt/median for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NIRFigGjsnGrVar(
  RegData,
  valgtVar,
  preprosess = 1,
  hentData = 0,
  valgtMaal = "Gjsn",
  minald = 0,
  maxald = 110,
  datoFra = "2011-01-01",
  datoTil = "3000-01-01",
  aar = 0,
  grType = 99,
  InnMaate = 99,
  dodInt = "",
  erMann = "",
  grVar = "ShNavn",
  medKI = 1,
  lagFig = 1,
  outfile = "",
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{valgtMaal}{'Med' = median. Alt annet gir gjennomsnitt 

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alders 
    \item liggetid: Liggetid 
    \item Nas: Skår for sykepleieraktiviteter. (Nursing Activities Score). Per døgn.
    \item NEMS: Skår for ressursbruk per opphold. (Nine Equivalents of Nursing Manpower Use Score)
    \item NEMS24: NEMS-skår per døgn. 
    \item respiratortid: Tid tilbrakt i respirator
    \item respiratortidNonInv: Respiratortid, ikke-invasiv
    \item respiratortidInvMoverf: Respiratortid, invasiv m/overf.
    \item respiratortidInvUoverf: Respiratortid, invasiv u/overf.
    \item SAPSII: Skår for alvorlighetsgrad av sykdom.  (Simplified Acute Physiology Score II)
    \item SMR: (SAPSII-estimert dødelighet) Standardisert mortalitetsratio (Gir annen figurtype)
    \item PIMdod: PIM-estimert dødelighet (Gir annen figurtype)
   }

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 110)}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{grType}{Gjør gruppeutvalg på sykehustype
1: lokal-/sentralsykehus
2: lokal-/sentralsykehus
3: regionsykehus
99: alle (standard)}

\item{InnMaate}{0: Elektivt, 
6: Akutt medisinsk, 
8: Akutt kirurgisk, 
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{dodInt}{Levende/død ut fra intensiv. 
0: i live, 
1: død,   
alle: standard (alle andre verdier)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{lagFig}{Angir om figur skal lages eller ikke 0-ikke lag, 1-lag}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}
}
\value{
Søylediagram med gjennomsnitt/median av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med gjennomsnitt/median
for hvert sykehus og kan ta inn ulike numeriske variable.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
