% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigGjsnPaaror.R
\name{NIRFigGjsnPaaror}
\alias{NIRFigGjsnPaaror}
\title{Funksjon for å tilrettelegge variable for beregning.}
\usage{
NIRFigGjsnPaaror(RegData, valgtVar, valgtMaal = "Gjsn", prePost = 2,
  grVar = "ShNavn", outfile = "")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{prePost}{Før (1) eller etter (2) intervensjon}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}
}
\value{
Definisjon av valgt variabel.
}
\description{
Funksjon for å tilrettelegge variable for beregning.
}
