% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigAndelerPrePost.R
\name{NIRFigPrePost}
\alias{NIRFigPrePost}
\title{Søylediagram som fordeling av valgt variabel, målt både før og etter hjerneslaget}
\usage{
NIRFigPrePost(RegData, valgtVar, datoFra = "2012-04-01",
  datoTil = "2050-12-31", minald = 0, maxald = 130, erMann = "",
  diagnose = "", innl4t = "", NIHSSinn = "", outfile = "", reshID,
  enhetsUtvalg = 1, preprosess = 1, hentData = 0)
}
\arguments{
\item{valgtVar}{Hvilken variabel som skal visualiseres
Boligforh: Boligforhold ved innleggelse og 3 mnd etter slaget
Bosituasjon: Bosituasjon ved innleggelse og 3 mnd etter slaget
Toalett: Toalettbesok før innleggelse og 3 mnd etter slaget 
Forflytning: Evne til forflytning før innleggelse og 3 mnd etter slaget}
}
\value{
Søylediagram som fordeling av valgt variabel, målt både før og etter hjerneslaget
}
\description{
Funksjon som genererer en figur med som viser endring i en variabels fordeling fra før til etter hjerneslaget.
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
