% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigPrePostPaaror.R
\name{NIRFigPrePostPaaror}
\alias{NIRFigPrePostPaaror}
\title{Søylediagram med fordeling før og ved intervensjon}
\usage{
NIRFigPrePostPaaror(
  RegData = 0,
  valgtVar,
  preprosess = 1,
  hentData = 0,
  reshID = 0,
  enhetsUtvalg = 0,
  datoFra = "2023-11-07",
  datoTil = Sys.Date(),
  aar = 0,
  startDatoIntervensjon = "2025-01-01",
  sluttDatoIntervensjon = Sys.Date(),
  minald = 0,
  maxald = 110,
  erMann = "",
  InnMaate = "",
  dodInt = "",
  nivaa = 0,
  outfile = "",
  lagFig = 1,
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Standard 2019-01-01.
Registeret inneholder registreringer f.o.m. 2011}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{aar}{Vis registreringer for opphold som startet i dette året}

\item{startDatoIntervensjon}{startDatoIntervensjon}

\item{sluttDatoIntervensjon}{sluttDatoIntervensjon}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 110)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{InnMaate}{0: Elektivt,
6: Akutt medisinsk,
8: Akutt kirurgisk,
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{dodInt}{Levende/død ut fra intensiv.
0: i live,
1: død,
alle: standard (alle andre verdier)}

\item{nivaa}{Intensivavdelingens "nivå": 1-5: 'Overvåk', 'Postop', 'Gen<50%','Gen>50%', 'Barn'
Flervalgsparameter}

\item{outfile}{Filbane for lagring av figur.}

\item{lagFig}{Angir om figur skal lages eller ikke 0-ikke lag, 1-lag}

\item{...}{Ytterligere argumenter, for eksempel user for logging. Se detaljer i NIRUtvalgEnh og NIRVarTilrettelegg.}
}
\value{
Søylediagram med fordeling før og ved intervensjon
}
\description{
Funksjon som genererer en fordelingsfigur for to tidsperioder som skal sammenlignes
}
\details{
Nytt skjema tatt i bruk 7.nov 2023
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
