% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigSoyler.R
\name{NIRFigSoyler}
\alias{NIRFigSoyler}
\title{IKKE PÅBEGYNT....!!!}
\usage{
NIRFigSoyler(RegData, FigDataParam = "", outfile)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{valgtMaal}{'Med' = median. Alt annet gir gjennomsnitt 

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alders 
    \item SMR: Standardisert mortalitetsratio (Gir annen figurtype)
    \item liggetid: Liggetid 
    \item Nas: Skår for sykepleieraktiviteter. (Nursing Activities Score). Per døgn.
    \item NEMS: Skår for ressursbruk per opphold. (Nine Equivalents of Nursing Manpower Use Score)
    \item NEMS24: NEMS-skår per døgn. 
    \item respiratortid: Tid tilbrakt i respirator
    \item SAPSII: Skår for alvorlighetsgrad av sykdom.  (Simplified Acute Physiology Score II)
   }

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}
}
\value{
Søylediagram med gjennomsnitt/median av valgt variabel for hvert sykehus
}
\description{
bLA, BAL
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}

