% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigTidAndel.R
\name{NIRFigTidAndel}
\alias{NIRFigTidAndel}
\title{Lager tidslinjeplott}
\usage{
NIRFigTidAndel(RegData, AggVerdier, AggTot = 0, Ngr,
  tittel = "mangler tittel", smltxt = "", N, retn = "H", yAkseTxt = "",
  utvalgTxt = "", grTypeTxt = "", tidtxt, varTxt = "", grtxt2 = "",
  hovedgrTxt = "", valgtMaal = "Andel", cexgr = 1, medSml = 0,
  fargepalett = "BlaaOff", xAkseTxt = "", medKI = 0, KImaal = NA,
  outfile = "")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{AggVerdier}{Liste med AggVerdier i hver kategori (aldersgruppe, sykehus, ...). 
AggVerdier$Hoved er enten andelene for hele utvalget i hver kategori eller f.eks. egen enhet
AggVerdier$Rest er andelene for gruppa det sammenlignes mot.}

\item{AggTot}{Aggregert verdi for hele gruppa (ett tall)}

\item{Ngr}{Antall i hver gruppe}

\item{tittel}{Figurtittel}

\item{smltxt}{Tekststreng som angir hva vi evt. sammenlikner med. Benyttes ogs? til ? angi 
gruppenavnet (i f.eks. tekstetikett) hvis det er gjort utvalg p? f.eks. en sykehustype.}

\item{N}{Totalantall angitt i N$Hoved og N$Rest}

\item{retn}{Horisontale eller Vertikale s?yler ('H', 'V')}

\item{yAkseTxt}{Tekst p? y-aksen}

\item{utvalgTxt}{Tekst som angir hvilke utvalg som er gjort}

\item{grTypeTxt}{Tekst som angir}

\item{grtxt2}{Evt. undertekst p? kategoriene (Eks 28pst)}

\item{hovedgrTxt}{Angivelse av hovedgruppe, eks. eget sykehus}

\item{valgtMaal}{'Med' = median. Alt annet gir gjennomsnitt}

\item{cexgr}{Skalering av gruppetekst (trenger mindre font n?r flere grupper)}

\item{medSml}{Angir om vi skal gj?re sammenlikning eller ikke (verdier: 1/0)}

\item{fargepalett}{'BlaaOff'}

\item{xAkseTxt}{tekst p? x-aksen}

\item{KImaal}{Kvalitetsindikatorm?l}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{grtxt}{Navn p? gruppene/kategoriene (Eks 40-50 ?r)}
}
\value{
S?ylediagram med gjennomsnitt/median av valgt variabel for hvert sykehus
}
\description{
Funksjonen genererer linjediagram basert p? data som leveres funksjonen. Linjediagrammenes utseende
kan variere for ulike datatyper. Forel?pig (2017-05-22) viser den utvikling over tid for andeler.
}
\details{
Detajer: Her b?r man liste opp hvilke variable funksjonen benytter.
}
