% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRGjsnGrVar.R
\name{NIRGjsnGrVar}
\alias{NIRGjsnGrVar}
\title{Søylediagram med gjennomsnitt/median for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NIRGjsnGrVar(RegData, valgtVar, preprosess = 1, hentData = 0,
  valgtMaal = "Gjsn", minald = 0, maxald = 130, datoFra = "2011-01-01",
  datoTil = "3000-01-01", aar = 0, grType = 99, InnMaate = 99,
  dodInt = "", erMann = "", grVar = "ShNavn", medKI = 1, lagFig = 1,
  outfile)
}
\arguments{
\item{valgtMaal}{'Med' = median. Alt annet gir gjennomsnitt 

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alders 
    \item SMR: Standardisert mortalitetsratio (Gir annen figurtype)
    \item liggetid: Liggetid 
    \item Nas: Skår for sykepleieraktiviteter. (Nursing Activities Score). Per døgn.
    \item NEMS: Skår for ressursbruk per opphold. (Nine Equivalents of Nursing Manpower Use Score)
    \item NEMS24: NEMS-skår per døgn. 
    \item respiratortid: Tid tilbrakt i respirator
    \item SAPSII: Skår for alvorlighetsgrad av sykdom.  (Simplified Acute Physiology Score II)
   }

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}
}
\value{
Søylediagram med gjennomsnitt/median av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med gjennomsnitt/median
for hvert sykehus og kan ta inn ulike numeriske variable.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
