% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRGjsnTid.R
\name{NIRGjsnTid}
\alias{NIRGjsnTid}
\title{Utvikling over tid for gjennomsnitt/median av valgt variabel}
\usage{
NIRGjsnTid(RegData, valgtVar, datoFra = "2011-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, erMann = "", reshID,
  InnMaate = "", dodInt = "", tittel = 1, outfile = "",
  enhetsUtvalg = 1, valgtMaal = "", preprosess = 1, hentData = 0)
}
\arguments{
\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{valgtMaal}{'Gjsn': gir middelverdi (standard)
'Med': gir median}
}
\value{
Linjediagram som viser utvikling over tid for valgt variabel
}
\description{
Figuren viser gjennomsnitt/median per år med konfidensintervall for valgt variabel.
I bakgrunn vises konfidensintervall for det man har valt å sammenlikne med.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alder
    \item liggetid: Liggetid 
    \item respiratortid: Tid tilbrakt i respirator
    \item SAPSII: Skår for alvorlighetsgrad av sykdom.  (Simplified Acute Physiology Score II)
   }
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
   \itemize{
    \item 0: Hele landet
    \item 1: Egen enhet mot resten av landet (Standard)
    \item 2: Egen enhet
    \item 3: Egen enhet mot egen sykehustype
    \item 4: Egen sykehustype
    \item 5: Egen sykehustype mot resten av landet
    \item 6: Egen enhet mot egen region [NB: Intensivregiisteret mangler pt. variabel for region]
    \item 7: Egen region [NB: Mangler pt. variabel for region]
   \item 8: Egen region mot resten [NB: Mangler pt. variabel for region]
   	}
}
