% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRpaarorDataSQL.R
\name{NIRpaarorDataSQL}
\alias{NIRpaarorDataSQL}
\title{Henter data fra pårørendeskjema registrert for Intensiv og kobler til 
noen variabler fra hovedskjema.}
\usage{
NIRpaarorDataSQL(datoFra = "2015-12-01", datoTil = Sys.Date(), medH = 0)
}
\arguments{
\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Standard 2019-01-01.
Registeret inneholder registreringer f.o.m. 2011}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{medH}{kobler på variabler fra hovedskjema}
}
\value{
Henter dataramma RegData for Intensivregisteret
}
\description{
Henter data for Intensivregisterets database
}
