% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRPreprossesser.R
\name{NIRPreprosess}
\alias{NIRPreprosess}
\title{Preprosesser data fra Intensivregisteret}
\usage{
NIRPreprosess(RegData = RegData, lagreKvalIndData = 0)
}
\value{
Data En liste med det filtrerte datasettet (og sykehusnavnet som tilsvarer reshID, ikke pt)
}
\description{
Denne funksjonen navner om variabler og beregner evt. nye.
Funksjonen lager også et "offentlig" datasett som kan benyttes til beregning
av kvalitetsindikatorer og som kan legges ved pakken
}
