% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRRegDataSQL.R
\name{NIRRegDataSQL}
\alias{NIRRegDataSQL}
\title{Henter data registrert for Intensiv}
\usage{
NIRRegDataSQL(datoFra = "2011-01-01", datoTil = "2099-01-01",
  session = "")
}
\arguments{
\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}
}
\value{
Henter dataramma RegData for Intensivregisteret
}
\description{
Henter data for Intensivregisteret fra "staging"
}
