% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{NIRtabeller}
\alias{NIRtabeller}
\alias{tabAntOpphSh12mnd}
\title{Funksjoner for å lage tabeller Group of functions page title}
\usage{
tabAntOpphSh12mnd(RegData, datoTil)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato. Brukes}

\item{personIDvar}{Variabelen som angir pasientidentifikasjon}
}
\value{
Div tabeller
}
\description{
Fil som beregner div tabeller.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.
}
\section{Finne reinnleggelser After function section}{

Despite its location, this actually comes after the function section.
Fil som inneholder funksjoner for å lage tabeller, i første rekke tellinger av personer 
Aktuelle tabeller:
-Belegg (samlerapport), antall opphold per år og enhet, ant. pasienter per år og enhet, ant opph. per måned og enhet.
-tabAntOpphSh12mnd: Antall opphold per måned og enhet siste 12 måneder fram til datoTil. 
RegData må inneholde InnDato.
-tabAntOpphSh5Aar:Antall opphold per år og enhet siste 5 år (inkl. inneværende år) fram til datoTil. 
RegData må inneholde Aar.
}

\section{Antall opphold per enhet og måned}{

Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.
}

