% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRUtvalg.R
\name{NIRUtvalg}
\alias{NIRUtvalg}
\title{Funksjon som gjør utvalg av dataene, returnerer det filtrerte datasettet og utvalgsteksten.}
\usage{
NIRUtvalg(RegData, datoFra, datoTil, minald = 0, maxald = 130,
  erMann = "", InnMaate = "", grType = 99, overfPas = 99, dodInt = "",
  fargepalett = "BlaaOff")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{InnMaate}{0: Elektivt, 
6: Akutt medisinsk, 
8: Akutt kirurgisk, 
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{overfPas}{Overført under pågående intensivbehandling? 
1 = Pasienten er ikke overført
2 = Pasienten er overført}

\item{dodInt}{Levende/død ut fra intensiv. 
0: i live, 
1: død,   
alle, standard (alle andre verdier)}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Standard: BlaaRapp)}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det filtrerte datasettet og utvalgsteksten.
}
