% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRUtvalgEnh.R
\name{NIRUtvalgEnh}
\alias{NIRUtvalgEnh}
\title{Funksjon som gjør utvalg av dataene, returnerer det filtrerte datasettet og utvalgsteksten.}
\usage{
NIRUtvalgEnh(RegData, datoFra = 0, datoTil = 0, minald = 0,
  maxald = 130, erMann = "", InnMaate = "", aar = 0, grType = 99,
  enhetsUtvalg = 0, reshID = 0, dodInt = "", fargepalett = "BlaaOff")
}
\arguments{
\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Standard: BlaaRapp)}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det filtrerte datasettet og utvalgsteksten.
}
