% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRUtvalgOff.R
\name{NIRUtvalgOff}
\alias{NIRUtvalgOff}
\title{Funksjon som gjør utvalg av anonymiserte data, returnerer det filtrerte datasettet og utvalgsteksten.}
\usage{
NIRUtvalgOff(RegData, aldGr = 0, erMann = "", InnMaate = "", aar = 0,
  grType = 99, fargepalett = "BlaaOff")
}
\arguments{
\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Standard: BlaaRapp)}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Funksjon som gjør utvalg av anonymiserte data, returnerer det filtrerte datasettet og utvalgsteksten.
}
