% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRVarTilretteleggPaaror.R
\name{NIRVarTilretteleggPaaror}
\alias{NIRVarTilretteleggPaaror}
\title{Funksjon for å tilrettelegge variable for beregning.}
\usage{
NIRVarTilretteleggPaaror(RegData, valgtVar, grVar = "ShNavn",
  figurtype = "andeler")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{figurtype}{Hvilken figurtype det skal tilrettelegges variable for: 
'andeler', 'andelGrVar', 'andelTid', 'gjsnGrVar', 'gjsnTid'}
}
\value{
Definisjon av valgt variabel.
}
\description{
Denne funksjonen gjør utvalg og tilrettelegger variable (gitt ved valgtVar) til videre bruk. 
Videre bruk kan eksempelvis være beregning av AggVerdier eller gjennomsnitt. 
Funksjonen gjør også filtreringer som å fjerne ugyldige verdier for den valgte variabelen, samt ta høyde for avhengigheter med
andre variable. Det er også her man angir aksetekster og titler for den valgte variabelen. 
Her kan mye hentes til analysebok
}
\details{
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
   \itemize{
    \item 0: Hele landet
    \item 1: Egen enhet mot resten av landet (Standard)
    \item 2: Egen enhet
    \item 3: Egen enhet mot egen sykehustype
    \item 4: Egen sykehustype
    \item 5: Egen sykehustype mot resten av landet
    \item 6: Egen enhet mot egen region [NB: Intensivregiisteret mangler pt. variabel for region]
    \item 7: Egen region [NB: Mangler pt. variabel for region]
   \item 8: Egen region mot resten [NB: Mangler pt. variabel for region]
   	}
}
