% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{oppsumFerdigeRegTab}
\alias{oppsumFerdigeRegTab}
\title{Ferdigstilte registreringer}
\usage{
oppsumFerdigeRegTab(
  RegData,
  valgtRHF = "Alle",
  datoFra = "2020-01-01",
  datoTil = Sys.Date(),
  bekr = 9,
  erMann = 9,
  resp = 9,
  dodInt = 9
)
}
\arguments{
\item{RegData}{beredskapsskjema}
}
\value{
nøkkeltalltabell for ferdigstilte registreringer
}
\description{
Ferdigstilte registreringer
}
