% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportObjects.R
\name{readmission72hours}
\alias{readmission72hours}
\title{Provider of report objects}
\usage{
readmission72hours()
}
\arguments{
\item{Tittel}{String Title to show in empty chart}

\item{infoText}{String Text to shown in empty chart}
}
\value{
plotObj Highchart object

tableObj Table object
}
\description{
Provide report object fractions of readmittance within 72 hours
}

