% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRRegData01Off.R
\name{RegData01Off}
\alias{RegData01Off}
\title{Lager "anonymt" minimumsdatasett basert på rådata for å beregne kvalitetsindikatorer.}
\usage{
RegData01Off(RegData, valgtVar, datoFra = "2016-01-01",
  datoTil = "3000-01-01", tilleggsVar = 0, hentData = 0, rand = 0)
}
\arguments{
\item{tilleggsVar}{Variable som benyttes til filtrering eller gruppering av valgt variabel (valgtVar)
Aktuelle valg: Aar, erMann, ShNavn, Mnd,}

\item{rand}{0-"ekte data", 1-tilfeldig resultat.}
}
\value{
Definisjon av valgt variabel.
}
\description{
Datasettet som lagres inneholder utvalgs-/filtreringsvariable, samt teller (og nevner) for 
kvalitetsindikatoren som skal beregnes. Både teller og nevner er definert som en indikatorvariabel
hvor 1 angir om rada/hendelsen skal inkluderes i beregning av kvalitetsindikatoren eller ikke.
}
\details{
Funksjonen benytter funksjonene: NIRRegDataSQL, NIRPreprosess, NIRVarTilrettelegg, NIRUtvalgEnh
og NIRFigSoyler

Argumentet \emph{valgtVar} angir hvilken kvalitetsindikator har følgende valgmuligheter:
   \itemize{
    \item alder_u18: Pasienter under 18 år 
    \item alder_over80: Pasienter over 80 år (>=80)
    \item dod30d: Pasienter som dør innen 30 dager etter innleggelse
    \item dodeIntensiv: Pasienter som dør på intensivavdelinga. 
    \item innMaate: Hastegrad inn på intensiv (Elektivt, Akutt medisinsk, Akutt kirurgisk)
	Dette valget viser en annen figurtype.
    \item respStotte: Pasienter som har fått respiratorstøtte
    \item reinn: Andel reinnlagte (fjerner ukjente) Kvalitetsindikator
   }
}
