% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{SentralmaalTab}
\alias{SentralmaalTab}
\title{Tabell med sentralmål, min,maks IQR}
\usage{
SentralmaalTab(
  RegData,
  datoFra = "2020-01-01",
  datoTil = Sys.Date(),
  erMann = 9
)
}
\arguments{
\item{RegData}{beredskapsskjema}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Standard 2019-01-01.
Registeret inneholder registreringer f.o.m. 2011}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}
}
\value{
tabell, sentralmål
}
\description{
Tabell med sentralmål, min,maks IQR
}
