% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appIntensiv.R
\name{server_intensiv}
\alias{server_intensiv}
\title{Serverdek til Intensiv-appen}
\usage{
server_intensiv(input, output, session)
}
\arguments{
\item{input}{Shiny input-objekt}

\item{output}{Shiny output-objekt}

\item{session}{Shiny session-objekt}
}
\value{
Brukergrensesnittet (ui) til intensiv-appen
}
\description{
Serverdek til Intensiv-appen
}
