% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{SorterOgNavngiTidsEnhet}
\alias{SorterOgNavngiTidsEnhet}
\title{Tilrettelegge tidsenhetvariabel:
Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.}
\usage{
SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Aar", tab = 0)
}
\description{
Tilrettelegge tidsenhetvariabel:
Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.
}
