% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%m-\%}
\alias{\%m-\%}
\title{Subtract months to a date without exceeding the last day of the new month}
\usage{
e1 \%m-\% e2 e1
}
\description{
Subtract months to a date without exceeding the last day of the new month
}
\keyword{internal}
