% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R, R/NIRtabellerBeredskap.R
\name{TabAlder}
\alias{TabAlder}
\title{Aldersfordeling, tabell}
\usage{
TabAlder(
  RegData,
  reshID = 0,
  enhetsNivaa = "Alle",
  skjemastatus = 9,
  resp = 9,
  bekr = 9,
  dodInt = 9,
  erMann = 9,
  sens = 0
)

TabAlder(
  RegData,
  reshID = 0,
  enhetsNivaa = "Alle",
  skjemastatus = 9,
  resp = 9,
  bekr = 9,
  dodInt = 9,
  erMann = 9,
  sens = 0
)
}
\arguments{
\item{RegData}{datatabell, beredskapsdata}

\item{reshID}{avdelingsresh}

\item{enhetsNivaa}{enhetsnivå}

\item{dodInt}{Levende/død ut fra intensiv.
0: i live,
1: død,
alle: standard (alle andre verdier)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{sens}{maskere celler <3. 0-nei, 1-ja}
}
\value{
aldersfordeling

aldersfordeling
}
\description{
Aldersfordeling, tabell

Aldersfordeling, tabell
}
\examples{
\dontrun{
}
\dontrun{
TabAlder(RegData=CoroData, enhetsNivaa='HF')
}
}
