% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{TabAlderGml}
\alias{TabAlderGml}
\title{Aldersfordeling, tabell}
\usage{
TabAlderGml(
  RegData,
  valgtRHF = "Alle",
  skjemastatus = 9,
  resp = 9,
  bekr = 9,
  dodInt = 9,
  erMann = 9
)
}
\arguments{
\item{RegData}{datatabell, beredskapsdata}
}
\value{
aldersfordeling, gruppert
}
\description{
Aldersfordeling, tabell
}
\examples{
\dontrun{
TabAlder(RegData=CoroData, enhetsNivaa='HF')
}
}
