% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{tabAntOpphPasSh5Aar}
\alias{tabAntOpphPasSh5Aar}
\title{Antall registreringer/pasienter siste 5 år:}
\usage{
tabAntOpphPasSh5Aar(RegData, gr = "opph", datoTil)
}
\arguments{
\item{RegData}{data}

\item{gr}{gruppering opphold (opph, standard), pasienter (pas)}

\item{datoTil}{sluttdato}
}
\description{
Antall registreringer/pasienter siste 5 år:
}
