% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{tabAntOpphSh5Aar}
\alias{tabAntOpphSh5Aar}
\title{Antall opphold siste 5 år}
\usage{
tabAntOpphSh5Aar(RegData, datoTil)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato}
}
\description{
Antall opphold siste 5 år
}
