% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{tabAntOpphShAar}
\alias{tabAntOpphShAar}
\title{Antall opphold siste år}
\usage{
tabAntOpphShAar(RegData, datoTil, antAar = 5)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato}

\item{antAar}{antall år som skal vises}
}
\description{
Antall opphold siste år
}
