% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{tabBelegg}
\alias{tabBelegg}
\title{Funksjoner for å lage tabeller}
\usage{
tabBelegg(RegData, tidsenhet = "Aar", datoTil, enhetsUtvalg = 0, reshID = 0)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato. Brukes i tabellene AntOpph per 12 mnd og Belegg}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{personIDvar}{Variabelen som angir pasientidentifikasjon}
}
\description{
Belegg (antall opphold, pasienter og intensivdøgn)
Siste inntil 5 år eller siste inntil 12 måneder/kvartal/halvår
}
