% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{tabNokkeltallNord}
\alias{tabNokkeltallNord}
\title{Nøkkeltall (antall opph., pasienter,  intensivdøgn, samt div oversiktstall)}
\usage{
tabNokkeltallNord(RegData, tidsenhet = "Aar", sykehus = "Alle")
}
\arguments{
\item{RegData}{dataramme}

\item{sykehus}{HelseenhetKortnavn eller HF-navn}
}
\description{
Nøkkeltall (antall opph., pasienter,  intensivdøgn, samt div oversiktstall)
}
