% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{tabNokkeltallUtvid}
\alias{tabNokkeltallUtvid}
\title{Nøkkeltall (antall opph., pasienter,  intensivdøgn, samt div oversiktstall)}
\usage{
tabNokkeltallUtvid(
  RegData,
  tidsenhet = "Aar",
  sykehus = "Alle",
  datoFra = "2016-01-01",
  datoTil = Sys.Date()
)
}
\arguments{
\item{RegData}{dataramme}

\item{sykehus}{HelseenhetKortnavn eller HF-navn}
}
\description{
Nøkkeltall (antall opph., pasienter,  intensivdøgn, samt div oversiktstall)
}
