% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{TabOppsumLuftvei}
\alias{TabOppsumLuftvei}
\title{Oppsummeringer, luftveisinfeksjonspasienter}
\usage{
TabOppsumLuftvei(RegData, pgaLuftvei = 0)
}
\arguments{
\item{RegData}{}
}
\value{
oppsummering i tabellform
}
\description{
Oppsummeringer, luftveisinfeksjonspasienter
}
