% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{tabOverforinger}
\alias{tabOverforinger}
\title{Finne overføringer til/fra en enhet}
\usage{
tabOverforinger(
  RegData,
  datoFra = Sys.Date() - 365,
  datoTil = Sys.Date(),
  reshID = 0,
  velgAvd = 0,
  enhetsUtvalg = 2,
  overfFraSh = 1
)
}
\arguments{
\item{RegData}{data}

\item{datoFra}{startdato}

\item{datoTil}{sluttdato. Brukes i tabellene AntOpph per 12 mnd og Belegg}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{}

\item{overfFraSh}{- overføring fra (1) eller til (0) den aktuelle enheten}
}
\description{
Finne overføringer til/fra en enhet
}
