% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{tellInfluensa}
\alias{tellInfluensa}
\title{Telle antall influensaskjema for en gitt periode}
\usage{
tellInfluensa(datoFra = "2020-09-01", datoTil = Sys.Date(), reshID = 0)
}
\arguments{
\item{datoFra}{fra og med dato}

\item{datoTil}{til og med dato}

\item{reshID}{resh for aktuell enhet}
}
\value{
Tekst med ant. influensaskjema
}
\description{
Telle antall influensaskjema for en gitt periode
}
