% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{tilretteleggKvalIndData}
\alias{tilretteleggKvalIndData}
\title{Funksjon for å tilrettelegge kvalitetsindikatordata. Data skal  
være på forma 0-1 for "andelsvariabler". For indikatorer som baserers på gjennomsnitt/median,
må faktiske observasjoner benyttes.}
\usage{
tilretteleggKvalIndData(RegData, datoFra = "2016-01-01", datoTil = Sys.Date())
}
\arguments{
\item{RegData}{dataramme}

\item{datoFra}{startdato}

\item{datoTil}{sluttdato}
}
\value{

}
\description{
Funksjon for å tilrettelegge kvalitetsindikatordata. Data skal  
være på forma 0-1 for "andelsvariabler". For indikatorer som baserers på gjennomsnitt/median,
må faktiske observasjoner benyttes.
}
