% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appIntensiv.R
\name{ui_intensiv}
\alias{ui_intensiv}
\title{Brukergrensesnitt (ui) til Intensiv-appen}
\usage{
ui_intensiv()
}
\value{
Brukergrensesnittet (ui) til intensiv-appen
}
\description{
Brukergrensesnitt (ui) til Intensiv-appen
}
