% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiInputModule.R
\name{uiInputModule}
\alias{uiInputModule}
\title{Modulised UI for shiny}
\usage{
uiInputModule(id, label = "Brukervalg")
}
\description{
uiInputModule provides common ui uc for shiny intensiv
}
\details{
@param id namespace for this module
 @param label Label for this module
 @export
}

