% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppNIRbrukervalg.R
\name{AppNIRbrukervalg}
\alias{AppNIRbrukervalg}
\title{Funksjon som inneholder alle brukervalg for Intensivregisteret}
\usage{
AppNIRbrukervalg(ark = "angi sideID")
}
\value{
Visning av brukervalg
}
\description{
Denne funksjonen beregner ikke noe. :-|
}
\details{
Mer info?

 @param id namespace for this module
 @param inputId id for brukervalget til den aktuelle figur/tabelltypen
 @param label Label for this module
}
