% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{dataTilOffVisning}
\alias{dataTilOffVisning}
\title{Generere data til interaktive nettsider SKDE}
\usage{
dataTilOffVisning(
  RegData = RegData,
  valgtVar,
  aar = 0,
  indID = "indDummy",
  filUt = "dummy"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{- beinsmLavPre, peropKompDura, sympVarighUtstr, p.t. 10 kvalitetsind.}

\item{indID}{indikator-id, eks. 'ind1', 'ind2', osv.}

\item{filUt}{tilnavn for utdatatabell (fjern?)}
}
\value{
Datafil til interaktive nettsider, SKDE
}
\description{
Generere data til interaktive nettsider SKDE
}
