% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{FinnReinnleggelser}
\alias{FinnReinnleggelser}
\title{Fil som inneholder hjelpefunksjoner.}
\usage{
FinnReinnleggelser(RegData, PasientID = "PasientID")
}
\arguments{
\item{RegData}{data}
}
\value{
Variabelen Reinn lagt til datasettet
}
\description{
FinnReinnleggelser beregner reinnleggelser fra DateAdmittedIntensive og DateDischargedIntensive
}
