% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{FinnReinnleggelser}
\alias{FinnReinnleggelser}
\title{Finne reinnleggelser}
\usage{
FinnReinnleggelser(RegData, PasientID = "PasientID")
}
\arguments{
\item{RegData}{dataramme}

\item{PasientID}{pasientid}
}
\value{
Dataramme med reinnleggelser
}
\description{
Finne reinnleggelser
}
