% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{FinnReinnleggelser}
\alias{FinnReinnleggelser}
\title{Fil med div hjelpefunksjoner.Group of functions Description section}
\usage{
FinnReinnleggelser(RegData, PasientID = "PasientID")
}
\arguments{
\item{RegData}{data}

\item{PasientID}{Variabelen som angir pasientidentifikasjon}
}
\value{
Div hjelpefunksjoner
}
\description{
Detaljer. kommer senereGroup of functions Details paragraph.
}
\details{
Fil som inneholder hjelpefunksjoner. 
FinnReinnleggelser beregner reinnleggelser fra DateAdmittedIntensive og DateDischargedIntensive
SorterOgNavngiTidsEnhet Legger til tidsenhetene Aar, Halvaar, Mnd og Kvartal
}
