% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{hjelpeFunksjoner}
\alias{hjelpeFunksjoner}
\alias{FinnReinnleggelser}
\alias{SorterOgNavngiTidsEnhet}
\alias{lageTulleData}
\title{Hjelpefunksjoner. Group of functions page title}
\usage{
FinnReinnleggelser(RegData, PasientID = "PasientID")

SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Aar")

lageTulleData(RegData, varBort = "ShNavn", antSh = 27)
}
\arguments{
\item{RegData}{data}

\item{PasientID}{Variabelen som angir pasientidentifikasjon}
}
\value{
Div hjelpefunksjoner
}
\description{
Fil med div hjelpefunksjoner.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.
}
\section{Finne reinnleggelser After function section}{

Despite its location, this actually comes after the function section.
Fil som inneholder hjelpefunksjoner. 
FinnReinnleggelser beregner reinnleggelser fra DateAdmittedIntensive og DateDischargedIntensive
SorterOgNavngiTidsEnhet Legger til tidsenhetene Aar, Halvaar, Mnd og Kvartal
}

\section{Tilrettelegge tidsenhetvariabel}{

Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.
}

