% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRAndelerGrVar.R
\name{NIRAndelerGrVar}
\alias{NIRAndelerGrVar}
\title{Søylediagram med AggVerdier for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NIRAndelerGrVar(RegData, valgtVar, datoFra = 0, datoTil = 0, aar = 0,
  minald = 0, maxald = 130, aldGr = 0, grType = 99, grVar = "",
  InnMaate = 99, dodInt = "", erMann = "", hentData = 0,
  utvalgsInfo = "", tittel = "", sortAvtagende = TRUE, preprosess = 1,
  outfile = "", lagFig = 1, offData = 0)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{grType}{Gjør gruppeutvalg på sykehustype
1: lokal-/sentralsykehus
2: lokal-/sentralsykehus
3: regionsykehus
99: alle (standard)}

\item{InnMaate}{0: Elektivt, 
6: Akutt medisinsk, 
8: Akutt kirurgisk, 
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{dodInt}{Levende/død ut fra intensiv. 
0: i live, 
1: død,   
alle: standard (alle andre verdier)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{lagFig}{Angir om figur skal lages eller ikke 0-ikke lag, 1-lag}

\item{aldGr:}{Aldersgrupper. Brukes i offentliggjøringsfigurer}

\item{tittel:}{Hvis vil angi tittel direkte}

\item{utvalgsInfo:}{Hvis datafil lagret med utvalgsinfo}

\item{offData:}{Hvis vi benytter anonymiserte 01-data til offentliggjøring.}

\item{sortAvtagende:}{sortere søylene i figuren avtagende eller stigende.}
}
\value{
Søylediagram med AggVerdier av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med AggVerdier av en gitt variabel for ei valgt gruppering, 
f.eks. enheter. I øyeblikket benytter funksjonen bare 'ShNavn' som grupperingsvariabel, men 
andre valg kan lett inkluderes. Funksjonen er tilrettelagt for også å kunne benytte "01-data", 
dvs. kodede, anonymiserte data som benyttes til offentlige kvalitetsindikatorer.
}
\details{
Funksjonen benytter funksjonene: NIRRegDataSQL, NIRPreprosess, NIRVarTilrettelegg, NIRUtvalgEnh
og NIRFigSoyler.
Figurtypen som genereres er avhengig av valgtVar. Ved valgtVar='InnMaate', får man ei oversikt oversikt
over andel av innkomsttype ved hver enhet, dvs. ei fordeling av de tre innkomsttypene for hver enhet.
For andre "valgtVar" viser figuren andel av den valgte variabelen for hver enhet.
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder_u18: Pasienter under 18 år 
    \item alder_over80: Pasienter over 80 år (>=80)
    \item dod30d: Pasienter som dør innen 30 dager etter innleggelse
    \item dodeIntensiv: Pasienter som dør på intensivavdelinga. 
    \item innMaate: Hastegrad inn på intensiv (Elektivt, Akutt medisinsk, Akutt kirurgisk)
	Dette valget viser en annen figurtype.
    \item respStotte: Pasienter som har fått respiratorstøtte
    \item reinn: Andel reinnlagte (fjerner ukjente) Kvalitetsindikator
   }
}
