% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigAndelTid.R
\name{NIRFigAndelTid}
\alias{NIRFigAndelTid}
\title{Tidstrend av andel pasienter}
\usage{
NIRFigAndelTid(RegData, valgtVar, datoFra = "2011-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, erMann = "",
  InnMaate = "", dodInt = "", reshID, outfile = "", enhetsUtvalg = 1,
  preprosess = 1, hentData = 0)
}
\value{
Figur som viser tidstrend, dvs. andel av valgt variabel for hvert år.
}
\description{
Denne funksjonen lager et søylediagram som viser andeler (fordeling) av valgt variabel
filtrert på de utvalg som er gjort.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder_u18: Pasienter under 18 år 
    \item alder_over80: Pasienter over 80 år  (>=80)
    \item dod30d: Pasienter som dør innen 30 dager etter innleggelse #ut:dodeSykehus
    \item dodeIntensiv: Pasienter som dør på intensivavdelinga. 
 \item liggetidDod: Andel av total liggetid brukt på de som dør på intensiv
    \item respiratortidDod: Respiratortid brukt på de som dør på intensiv
    \item respStotte: Pasienter som har fått respiratorstøtte
    \item reinn: Andel reinnlagte (kun hvor dette er registrert, dvs. fjerner ukjente)
   }
}
