% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigAndelTid.R
\name{NIRFigAndelTid}
\alias{NIRFigAndelTid}
\title{Tidstrend av andel opphold}
\usage{
NIRFigAndelTid(
  RegData,
  valgtVar = "alder_u18",
  datoFra = "2011-01-01",
  datoTil = Sys.Date(),
  tidsenhet = "Aar",
  minald = 0,
  maxald = 110,
  erMann = "",
  InnMaate = "",
  dodInt = "",
  velgDiag = 0,
  reshID = 0,
  outfile = "",
  enhetsUtvalg = 0,
  preprosess = 1,
  hentData = 0,
  lagFig = 1,
  offData = 0,
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Standard 2019-01-01.
Registeret inneholder registreringer f.o.m. 2011}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{tidsenhet}{Oppløsning på tidsaksen. Verdier: 'Aar' (standard), 'Halvaar', 'Kvartal','Mnd'}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 110)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{InnMaate}{0: Elektivt,
6: Akutt medisinsk,
8: Akutt kirurgisk,
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{dodInt}{Levende/død ut fra intensiv.
0: i live,
1: død,
alle: standard (alle andre verdier)}

\item{velgDiag}{Velge diagnose. 0-alle, 1-Bekreftede Covid-pasienter, 2-alle unntatt Covid-pasienter}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{lagFig}{Angir om figur skal lages eller ikke 0-ikke lag, 1-lag}
}
\value{
Figur som viser tidstrend, dvs. andel av valgt variabel for hvert år.
}
\description{
Denne funksjonen lager et linjediagram som viser utvikling over tid  for andeler av valgt variabel,
filtrert på de utvalg som er gjort.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder_u18: Pasienter under 18 år
    \item alder_over80: Pasienter over 80 år (>=80)
    \item dod30d: Pasienter som dør innen 30 dager etter innleggelse
    \item dodeIntensiv: Pasienter som dør på intensivavdelinga.
    \item ExtendedHemodynamicMonitoring: Utvidet hemodyn. overvåkning
 \item isolering: Isolasjon av pasient
 \item liggetidDod: Liggetid, døde
 \item nyreBeh: Nyrebehandling
    \item reinn: Andel reinnlagte (fjerner ukjente) Kvalitetsindikator
    \item respStotte: Pasienter som har fått respiratorstøtte
    \item respiratortidInvMoverf: Respiratortid, inv. < 2,5d m/overføringer
    \item respiratortidInvUoverf: Respiratortid, inv. < 2,5d u/overføringer
    \item respiratortidDod: Respiratortid, døde
    \item Trakeostomi: Andel som har fått trakeostomi (kat 2 og 3)
    \item trakAapen: Trakeostomi, åpen
   }
}
