% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigGjsnGrVar.R, R/NIRGjsnGrVar.R
\name{NIRFigGjsnGrVar}
\alias{NIRFigGjsnGrVar}
\title{Søylediagram med gjennomsnitt/median for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NIRFigGjsnGrVar(RegData, valgtVar, valgtMaal = "Gjsn", minald = 0,
  maxald = 130, datoFra = "2011-01-01", datoTil = "3000-01-01",
  grType = 99, InnMaate = 99, dodInt = "", erMann = "",
  preprosess = 1, hentData = 0, grVar = "ShNavn", lagFig = 1, outfile)

NIRFigGjsnGrVar(RegData, valgtVar, valgtMaal = "Gjsn", minald = 0,
  maxald = 130, datoFra = "2011-01-01", datoTil = "3000-01-01",
  grType = 99, InnMaate = 99, dodInt = "", erMann = "",
  preprosess = 1, hentData = 0, grVar = "ShNavn", lagFig = 1, outfile)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{valgtMaal}{'Med' = median. Alt annet gir gjennomsnitt 

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alders 
    \item SMR: Standardisert mortalitetsratio (Gir annen figurtype)
    \item liggetid: Liggetid 
    \item Nas: Skår for sykepleieraktiviteter. (Nursing Activities Score). Per døgn.
    \item NEMS: Skår for ressursbruk per opphold. (Nine Equivalents of Nursing Manpower Use Score)
    \item NEMS24: NEMS-skår per døgn. 
    \item respiratortid: Tid tilbrakt i respirator
    \item SAPSII: Skår for alvorlighetsgrad av sykdom.  (Simplified Acute Physiology Score II)
   }

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{grType}{Gjør gruppeutvalg for
 1 eller 2: lokal-/sentralsykehus
3: regionsykehus}

\item{InnMaate}{0: Elektivt, 
6: Akutt medisinsk, 
8: Akutt kirurgisk, 
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{dodInt}{Levende/død ut fra intensiv. 
0: i live, 
1: død,   
alle, standard (alle andre verdier)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{valgtMaal}{'Med' = median. Alt annet gir gjennomsnitt 

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alders 
    \item SMR: Standardisert mortalitetsratio (Gir annen figurtype)
    \item liggetid: Liggetid 
    \item Nas: Skår for sykepleieraktiviteter. (Nursing Activities Score). Per døgn.
    \item NEMS: Skår for ressursbruk per opphold. (Nine Equivalents of Nursing Manpower Use Score)
    \item NEMS24: NEMS-skår per døgn. 
    \item respiratortid: Tid tilbrakt i respirator
    \item SAPSII: Skår for alvorlighetsgrad av sykdom.  (Simplified Acute Physiology Score II)
   }

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}
}
\value{
Søylediagram med gjennomsnitt/median av valgt variabel for hvert sykehus

Søylediagram med gjennomsnitt/median av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med gjennomsnitt/median
for hvert sykehus og kan ta inn ulike numeriske variable.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus

Funksjon som genererer en figur med gjennomsnitt/median
for hvert sykehus og kan ta inn ulike numeriske variable.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...

Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}

