% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigGjsnTid.R
\name{NIRFigGjsnTid}
\alias{NIRFigGjsnTid}
\title{Utvikling over tid for gjennomsnitt/median av valgt variabel}
\usage{
NIRFigGjsnTid(RegData, valgtVar, datoFra = "2011-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, erMann = "", reshID,
  InnMaate = "", dodInt = "", tittel = 1, outfile = "",
  enhetsUtvalg = 1, valgtMaal = "", preprosess = 1, hentData = 0)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{InnMaate}{0: Elektivt, 
6: Akutt medisinsk, 
8: Akutt kirurgisk, 
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{dodInt}{Levende/død ut fra intensiv. 
0: i live, 
1: død,   
alle, standard (alle andre verdier)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{valgtMaal}{'Gjsn': gir middelverdi (standard)
'Med': gir median}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\value{
Linjediagram som viser utvikling over tid for valgt variabel
}
\description{
Figuren viser gjennomsnitt/median per år med konfidensintervall for valgt variabel.
I bakgrunn vises konfidensintervall for det man har valt å sammenlikne med.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alder
    \item liggetid: Liggetid 
    \item respiratortid: Tid tilbrakt i respirator
    \item SAPSII: Skår for alvorlighetsgrad av sykdom.  (Simplified Acute Physiology Score II)
   }
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
   \itemize{
    \item 0: Hele landet
    \item 1: Egen enhet mot resten av landet (Standard)
    \item 2: Egen enhet
    \item 3: Egen enhet mot egen sykehustype
    \item 4: Egen sykehustype
    \item 5: Egen sykehustype mot resten av landet
    \item 6: Egen enhet mot egen region [NB: Intensivregiisteret mangler pt. variabel for region]
    \item 7: Egen region [NB: Mangler pt. variabel for region]
   \item 8: Egen region mot resten [NB: Mangler pt. variabel for region]
   	}
}
