% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigInnMaate.R
\name{NIRFigInnMaate}
\alias{NIRFigInnMaate}
\title{Funksjon som genererer en figur med fordeling av innkomsttype
f.eks. enheter. I øyeblikket benytter funksjonen bare 'ShNavn' som grupperingsvariabel, men 
andre valg kan lett inkluderes. Funksjonen er tilrettelagt for også å kunne benytte "01-data", 
dvs. kodede, anonymiserte data som benyttes til offentlige kvalitetsindikatorer.}
\usage{
NIRFigInnMaate(RegData, valgtVar = "InnMaate", datoFra = "2010-01-01",
  datoTil = "3000-01-01", aar = 0, minald = 0, maxald = 130,
  grType = 99, grVar = "ShNavn", InnMaate = 99, dodInt = "",
  erMann = "", hentData = 0, preprosess = 1, outfile = "")
}
\arguments{
\item{tittel:}{Hvis vil angi tittel direkte}

\item{utvalgsInfo:}{Hvis datafil lagret med utvalgsinfo}
}
\value{
Søylediagram med AggVerdier av valgt variabel for hvert sykehus
}
\description{
Funksjonen benytter funksjonene: NIRRegDataSQL, NIRPreprosess, NIRVarTilrettelegg, NIRUtvalgEnh
og NIRFigSoyler.
Figurtypen som genereres er avhengig av valgtVar. Ved valgtVar='InnMaate', får man ei oversikt oversikt
over andel av innkomsttype ved hver enhet, dvs. ei fordeling av de tre innkomsttypene for hver enhet.
For andre "valgtVar" viser figuren andel av den valgte variabelen for hver enhet.
}
