% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigInnMaate.R
\name{NIRFigInnMaate}
\alias{NIRFigInnMaate}
\title{Funksjon som genererer en figur med fordeling av innkomsttype
f.eks. enheter. I øyeblikket benytter funksjonen bare 'ShNavn' som grupperingsvariabel, men
andre valg kan lett inkluderes. Funksjonen er tilrettelagt for også å kunne benytte "01-data",
dvs. kodede, anonymiserte data som benyttes til offentlige kvalitetsindikatorer.}
\usage{
NIRFigInnMaate(
  RegData,
  valgtVar = "InnMaate",
  datoFra = "2010-01-01",
  datoTil = "3000-01-01",
  aar = 0,
  minald = 0,
  maxald = 110,
  velgDiag = 0,
  grType = 99,
  grVar = "ShNavn",
  InnMaate = 99,
  dodInt = "",
  erMann = "",
  hentData = 0,
  preprosess = 1,
  outfile = "",
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Standard 2019-01-01. 
Registeret inneholder registreringer f.o.m. 2011}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 110)}

\item{velgDiag}{Velge diagnose. 0-alle, 1-Bekreftede Covid-pasienter}

\item{grType}{Gjør gruppeutvalg på sykehustype
1: lokal-/sentralsykehus
2: lokal-/sentralsykehus
3: regionsykehus
99: alle (standard)}

\item{InnMaate}{0: Elektivt, 
6: Akutt medisinsk, 
8: Akutt kirurgisk, 
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{dodInt}{Levende/død ut fra intensiv. 
0: i live, 
1: død,   
alle: standard (alle andre verdier)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{tittel:}{Hvis vil angi tittel direkte}

\item{utvalgsInfo:}{Hvis datafil lagret med utvalgsinfo}
}
\value{
Søylediagram med AggVerdier av valgt variabel for hvert sykehus
}
\description{
Funksjonen benytter funksjonene: NIRRegDataSQL, NIRPreprosess, NIRVarTilrettelegg, NIRUtvalgEnh
og NIRFigSoyler.
Figurtypen som genereres er avhengig av valgtVar. Ved valgtVar='InnMaate', får man ei oversikt oversikt
over andel av innkomsttype ved hver enhet, dvs. ei fordeling av de tre innkomsttypene for hver enhet.
For andre "valgtVar" viser figuren andel av den valgte variabelen for hver enhet.
}
