% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigPrePostPaaror.R
\name{NIRFigPrePostPaaror}
\alias{NIRFigPrePostPaaror}
\title{Søylediagram med fordeling før og ved intervensjon}
\usage{
NIRFigPrePostPaaror(RegData, valgtVar, outfile = "")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{valgtMaal}{'Med' = median. Alt annet gir gjennomsnitt 

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alders 
    \item SAPSII: Skår for alvorlighetsgrad av sykdom.  (Simplified Acute Physiology Score II)
   }

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}
}
\value{
Søylediagram med fordeling før og ved intervensjon
}
\description{
Funksjon som genererer en fordelingsfigur for to tidsperioder som skal sammenlignes
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
