% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigPrePostPaaror.R
\name{NIRFigPrePostPaaror}
\alias{NIRFigPrePostPaaror}
\title{Søylediagram med fordeling før og ved intervensjon}
\usage{
NIRFigPrePostPaaror(
  RegData = 0,
  valgtVar,
  datoFra = "2011-01-01",
  datoTil = "3000-12-31",
  aar = 0,
  startDatoIntervensjon = "2016-10-01",
  sluttDatoIntervensjon = Sys.Date(),
  minald = 0,
  maxald = 110,
  erMann = "",
  InnMaate = "",
  dodInt = "",
  outfile = "",
  grType = 99,
  preprosess = 1,
  hentData = 0,
  reshID = 0,
  enhetsUtvalg = 0,
  lagFig = 1,
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 110)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{InnMaate}{0: Elektivt, 
6: Akutt medisinsk, 
8: Akutt kirurgisk, 
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{dodInt}{Levende/død ut fra intensiv. 
0: i live, 
1: død,   
alle: standard (alle andre verdier)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{grType}{Gjør gruppeutvalg på sykehustype
1: lokal-/sentralsykehus
2: lokal-/sentralsykehus
3: regionsykehus
99: alle (standard)}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{lagFig}{Angir om figur skal lages eller ikke 0-ikke lag, 1-lag}

\item{valgtMaal}{'Med' = median. Alt annet gir gjennomsnitt 

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alders 
    \item SAPSII: Skår for alvorlighetsgrad av sykdom.  (Simplified Acute Physiology Score II)
   }

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}
}
\value{
Søylediagram med fordeling før og ved intervensjon
}
\description{
Funksjon som genererer en fordelingsfigur for to tidsperioder som skal sammenlignes
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
