% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRFigSoyler.R
\name{NIRFigSoyler}
\alias{NIRFigSoyler}
\title{Lager søylefigur}
\usage{
NIRFigSoyler(RegData, AggVerdier, AggTot = 0, Ngr,
  tittel = "mangler tittel", smltxt = "", N, retn = "H", yAkseTxt = "",
  utvalgTxt = "", grTypeTxt = "", soyletxt = "", grtxt, grtxt2 = "",
  hovedgrTxt = "", grVar = "", valgtMaal = "Andel", cexgr = 1,
  medSml = 0, fargepalett = "BlaaOff", xAkseTxt = "", medKI = 0,
  KImaal = NA, outfile = "")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{AggVerdier}{Liste med AggVerdier i hver kategori (aldersgruppe, sykehus, ...). 
AggVerdier$Hoved er enten andelene for hele utvalget i hver kategori eller f.eks. egen enhet
AggVerdier$Rest er andelene for gruppa det sammenlignes mot.}

\item{AggTot}{Aggregert verdi for hele gruppa (ett tall)}

\item{Ngr}{Antall i hver gruppe}

\item{tittel}{Figurtittel}

\item{smltxt}{Tekststreng som angir hva vi evt. sammenlikner med. Benyttes også til å angi 
gruppenavnet (i f.eks. tekstetikett) hvis det er gjort utvalg på f.eks. en sykehustype.}

\item{N}{Totalantall angitt i N$Hoved og N$Rest}

\item{retn}{Horisontale eller Vertikale søyler ('H', 'V')}

\item{yAkseTxt}{Tekst på y-aksen}

\item{utvalgTxt}{Tekst som angir hvilke utvalg som er gjort}

\item{grTypeTxt}{Tekst som angir}

\item{soyletxt}{Tekst PÅ hver søyle}

\item{grtxt}{Navn på gruppene/kategoriene (Eks 40-50 år)}

\item{grtxt2}{Evt. undertekst på kategoriene (Eks 28pst)}

\item{hovedgrTxt}{Angivelse av hovedgruppe, eks. eget sykehus}

\item{valgtMaal}{'Med' = median. Alt annet gir gjennomsnitt}

\item{cexgr}{Skalering av gruppetekst (trenger mindre font når flere grupper)}

\item{medSml}{Angir om vi skal gjøre sammenlikning eller ikke (verdier: 1/0)}

\item{fargepalett}{'BlaaOff'}

\item{xAkseTxt}{tekst på x-aksen}

\item{KImaal}{Kvalitetsindikatormål}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}
}
\value{
Søylediagram med gjennomsnitt/median av valgt variabel for hvert sykehus
}
\description{
Funksjonen genererer søylediagram basert på data som leveres funksjonen. Søylediagrammenes utseende
varierer noe for ulike datatyper, dvs. den kan vise både fordelingsfigur, AggVerdier eller sentralmål for 
gitt grupperingsvariabel, ...
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter.
}
