% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRPreprossesser.R
\name{NIRPreprosess}
\alias{NIRPreprosess}
\title{Preprosesser data fra Intensivregisteret}
\usage{
NIRPreprosess(RegData = RegData, reshID = reshID)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}
}
\value{
Data En liste med det filtrerte datasettet og sykehusnavnet som tilsvarer reshID
}
\description{
Denne funksjonen navner om variabler og beregner evt. nye.
}

