% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRPreprossesser.R
\name{NIRPreprosess}
\alias{NIRPreprosess}
\title{Preprosesser data fra Intensivregisteret}
\usage{
NIRPreprosess(RegData = RegData, skjema = 1)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{skjema}{hvilket skjema data som skal preprosesseres tilhører
1: hoved, 2: paaror, 3: influ, 4: beredsk (beredsk har egen preprosess-fil)}
}
\value{
Data En liste med det filtrerte datasettet (og sykehusnavnet som tilsvarer reshID, ikke pt)
}
\description{
Denne funksjonen navner om variabler og beregner evt. nye.
Funksjonen lager også et "offentlig" datasett som kan benyttes til beregning
av kvalitetsindikatorer og som kan legges ved pakken
}
