% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRUtvalgOff.R
\name{NIRUtvalgOff}
\alias{NIRUtvalgOff}
\title{Funksjon som gjør utvalg av anonymiserte data, returnerer det filtrerte datasettet og utvalgsteksten.}
\usage{
NIRUtvalgOff(
  RegData,
  aldGr = 0,
  erMann = "",
  InnMaate = "",
  aar = 0,
  grType = 99,
  fargepalett = "BlaaOff"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{InnMaate}{0: Elektivt,
6: Akutt medisinsk,
8: Akutt kirurgisk,
standard: alle (alt unntatt 0,6,8 / ikke spesifisert)}

\item{grType}{Gjør gruppeutvalg på sykehustype
1: lokal-/sentralsykehus
2: lokal-/sentralsykehus
3: regionsykehus
99: alle (standard)}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Standard: BlaaRapp)}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Funksjon som gjør utvalg av anonymiserte data, returnerer det filtrerte datasettet og utvalgsteksten.
}
