% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{SentralmaalTab}
\alias{SentralmaalTab}
\title{Tabell med sentralmål, min,maks IQR}
\usage{
SentralmaalTab(
  RegData,
  valgtRHF = "Alle",
  datoFra = "2020-01-01",
  datoTil = Sys.Date(),
  bekr = 9,
  erMann = 9,
  resp = 9,
  dodInt = 9
)
}
\arguments{
\item{RegData}{beredskapsskjema}
}
\value{
tabell, sentralmål
}
\description{
Tabell med sentralmål, min,maks IQR
}
