% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appIntensiv.R
\name{server_intensiv}
\alias{server_intensiv}
\title{Brukergrensesnitt (ui) til Intensiv-appen}
\usage{
server_intensiv(input, output, session)
}
\value{
Brukergrensesnittet (ui) til nger-appen
}
\description{
Brukergrensesnitt (ui) til Intensiv-appen
}
