% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/serverModule.R
\name{serverModule}
\alias{serverModule}
\title{Server logic function for shiny 'intensiv'}
\usage{
serverModule(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\value{
reportObjects list of objects returned from the module, typical
plotObj and tableObj
}
\description{
nakkeStandard is the module server function for shiny Intensiv
}

