% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRhjelpefunksjoner.R
\name{SorterOgNavngiTidsEnhet}
\alias{SorterOgNavngiTidsEnhet}
\title{Tilrettelegge tidsenhetvariabel:}
\usage{
SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Aar", tab = 0)
}
\arguments{
\item{RegData}{dataramme}

\item{tidsenhet}{'Aar' (standard), 'Halvaar', 'Kvartal', 'Mnd'}

\item{tab}{husker ikke hva denne gjør...}
}
\description{
Tilrettelegge tidsenhetvariabel:
}
