% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{statusECMOrespTab}
\alias{statusECMOrespTab}
\title{Antall som er  i ECMO/respirator MANGE ELEMENTER SOM IKKE KAN BENYTTES FOR INTENSIVSKJEMA}
\usage{
statusECMOrespTab(RegData, valgtRHF = "Alle", erMann = 9, luftvei = 0)
}
\arguments{
\item{RegData}{beredskapsskjema}
}
\value{
statustabell
}
\description{
Antall som er  i ECMO/respirator MANGE ELEMENTER SOM IKKE KAN BENYTTES FOR INTENSIVSKJEMA
}
