% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{tabAntOpphShMnd}
\alias{tabAntOpphShMnd}
\title{tabAntOpphShMnd antall opphold siste X (antMnd) mnd}
\usage{
tabAntOpphShMnd(
  RegData,
  datoTil = Sys.Date(),
  datoFra = "Ikke angitt",
  antMnd = 6
)
}
\arguments{
\item{RegData}{}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{antMnd}{antall måneder som skal vises}
}
\description{
tabAntOpphShMnd antall opphold siste X (antMnd) mnd
}
