% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabeller.R
\name{tabNokkeltall}
\alias{tabNokkeltall}
\title{Nøkkeltall (antall opph., pasienter,  intensivdøgn, samt div oversiktstall)}
\usage{
tabNokkeltall(
  RegData,
  tidsenhet = "Mnd",
  datoTil = Sys.Date(),
  enhetsUtvalg = 0,
  reshID = 0
)
}
\arguments{
\item{RegData}{dataramme}

\item{tidsenhet}{velg: Aar, Halvaar, Kvartal, Mnd (standard)}

\item{datoTil}{sluttdato}

\item{enhetsUtvalg}{enhetsutvalg}

\item{reshID}{enhetens resh-id}
}
\description{
Nøkkeltall (antall opph., pasienter,  intensivdøgn, samt div oversiktstall)
}
