% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{TabTidEnhet}
\alias{TabTidEnhet}
\title{Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.}
\usage{
TabTidEnhet(
  RegData,
  tidsenhet = "dag",
  erMann = 9,
  resp = 9,
  datoFra = 0,
  skjemastatus = 9,
  dodInt = 9,
  valgtRHF = "Alle",
  velgAvd = 0
)
}
\arguments{
\item{RegData}{dataramme med preprossesserte data}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned'}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{datoFra}{Vis hendelser fra og med dato}

\item{dodInt}{Levende/død ut fra intensiv.
0: i live,
1: død,
alle: standard (alle andre verdier)}

\item{enhetsNivaa}{'RHF', 'HF', 'ShNavn'}
}
\value{
Tabell med antall for hver tidsenhet og enhet MÅ OPPDATERES!
}
\description{
Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.
}
