% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/uiInputModule.R
\name{uiInputModule}
\alias{uiInputModule}
\title{Modulised UI for shiny}
\usage{
uiInputModule(id, label = "Brukervalg")
}
\arguments{
\item{id}{namespace for this module}

\item{label}{Label for this module}
}
\description{
uiInputModule provides common ui uc for shiny intensiv
}

